Updated company register fields and added tabs for influencer and business signup
