+++ client/src/pages/Login/BusinessLogin.tsx
@@ -40,9 +40,9 @@ export default function BusinessLogin() {
           margin="normal"
           required
           fullWidth
-          id="email"
-          label="Email Address"
-          name="email"
+          id="username-email"
+          label="Email Address/Username"
+          name="username-email"
           autoComplete="email"
           autoFocus
         />
