+++ lampsz/settings/local.py
@@ -12,3 +12,4 @@ SECRET_KEY = env(
 )
 # https://docs.djangoproject.com/en/dev/ref/settings/#allowed-hosts
 ALLOWED_HOSTS = env.list("ALLOWED_HOSTS", default=["localhost", "0.0.0.0", "127.0.0.1"])
+INSTALLED_APPS += ["whitenoise.runserver_nostatic"]
